Remove unused components and update HomeView
The `HelloWorld.vue`, `TheWelcome.vue`, and `WelcomeItem.vue` components have been completely removed from the project. The `HomeView.vue` file has been modified to remove the import and usage of the `TheWelcome` component. This cleanup likely reflects a restructuring or simplification of the project, removing introductory or example components that are no longer necessary.